requirements.txt
presidio_analyzer
presidio_anonymizer
spacy
datasets
PyYAML